more

+++ AI/Spike.cs
@@ -3,7 +3,7 @@
 public class Spike
 {
     public long ArrivalTime;
-    public byte Charge;
+    public double Charge;
     public Neuron Target;
     public Neuron? Source; //Some spikes originate from beyond the veil as background noise + energy injection.
 }
